Simplify datahandler classes by exploiting commonalities
